Move Channel object to InstrumentDesign
